init: get-refs
