+++ src/js/Keyboard.js
@@ -1,10 +1,11 @@
+import getRefs from './get-refs.js';
+const refs = getRefs();
+
 export default class Keyboard {
   constructor() {
     this.main = null;
     this.keysContainer = null;
     this.keys = [];
-    this.onInput = null;
-    this.onClose = null;
     this.value = '';
     this.capsLock = false;
   }
@@ -15,170 +16,13 @@ export default class Keyboard {
 
     this.main.classList.add('keyboard');
     this.keysContainer.classList.add('keyboard__keys');
-    this.keysContainer.appendChild(this._createKeys());
+    // this.keysContainer.appendChild(this._createKeys());
+
+    this.keys = this.keysContainer.querySelectorAll('.keyboard__key');
 
     this.main.appendChild(this.keysContainer);
     document.body.appendChild(this.main);
+
+    // refs.ourFriendsSlider.insertAdjacentHTML('beforeend', card.join(''));
   }
-
-  _createKeys() {
-    const fragment = document.createDocumentFragment();
-    const keyLayout = [
-      'backtick',
-      '1',
-      '2',
-      '3',
-      '4',
-      '5',
-      '6',
-      '7',
-      '8',
-      '9',
-      '0',
-      '-',
-      '=',
-      'backspace',
-      'tab',
-      'q',
-      'w',
-      'e',
-      'r',
-      't',
-      'y',
-      'u',
-      'i',
-      'o',
-      'p',
-      '[',
-      ']',
-      '\\',
-      'del',
-      'capslock',
-      'a',
-      's',
-      'd',
-      'f',
-      'g',
-      'h',
-      'j',
-      'k',
-      'l',
-      ';',
-      "'",
-      'enter',
-      'leftshift',
-      'z',
-      'x',
-      'c',
-      'v',
-      'b',
-      'n',
-      'm',
-      ',',
-      '.',
-      '/',
-      'uparrow',
-      'rightshift',
-      'leftctrl',
-      'win',
-      'leftalt',
-      'spacebar',
-      'rightalt',
-      'leftarrow',
-      'downarrow',
-      'rightarrow',
-      'rightctrl',
-    ];
-
-    const createIcon = iconName => {
-      return `<i class="icons">${iconName}</i>`;
-    };
-
-    keyLayout.forEach(key => {
-      const keyElement = document.createElement('button');
-      const insertLineBreack =
-        ['backspace', 'del', 'enter', 'shift', 'ctrl'].indexOf(key) !== -1;
-
-      keyElement.setAttribute('type', 'button');
-      keyElement.classList.add('keyboard__key');
-
-      switch (key) {
-        case 'backspace':
-          keyElement.classList.add('keyboard__key--wide');
-          keyElement.innerHTML = createIcon('backspase');
-
-          keyElement.addEventListener('click', () => {
-            this.value = this.value.substring(0, this.value.length - 1);
-            this._trigerEvent('onInput');
-          });
-          break;
-
-        case 'caps':
-          keyElement.classList.add(
-            'keyboard__key--wide',
-            'keyboard__key--activatable',
-          );
-          keyElement.innerHTML = createIcon('keyboard_capslock');
-
-          keyElement.addEventListener('click', () => {
-            this._toggleCapsLock();
-            keyElement.classList.toggle('keyboard__key--active', this.capsLock);
-          });
-          break;
-
-        case 'enter':
-          keyElement.classList.add('keyboard__key--wide');
-          keyElement.innerHTML = createIcon('keyboard_return');
-
-          keyElement.addEventListener('click', () => {
-            this.value += '\n';
-            this._trigerEvent('onInput');
-          });
-          break;
-
-        case 'spase':
-          keyElement.classList.add('keyboard__key--extra-wide');
-          keyElement.innerHTML = createIcon('space_bar');
-
-          keyElement.addEventListener('click', () => {
-            this.value += ' ';
-            this._trigerEvent('onInput');
-          });
-          break;
-
-        case 'done':
-          keyElement.classList.add(
-            'keyboard__key--wide',
-            'keyboard__key--dark',
-          );
-          keyElement.innerHTML = createIcon('space_bar');
-
-          keyElement.addEventListener('click', () => {
-            this.close();
-            this._trigerEvent('onClose');
-          });
-          break;
-        default:
-          keyElement.textContent = key.toLowerCase();
-
-          keyElement.addEventListener('click', () => {
-            this.value += this.capsLock ? key.toUpperCase() : key.toLowerCase();
-            this._trigerEvent('onInput');
-          });
-          break;
-      }
-
-      fragment.appendChild(keyElement);
-
-      if (insertLineBreack) {
-        fragment.appendChild(document.createElement('br'));
-      }
-    });
-
-    return fragment;
-  }
-
-  _trigerEvent() {}
-
-  render() {}
 }
