+++ src/js/get-refs.js
@@ -1,0 +1,5 @@
+export default function getRefs() {
+  return {
+    body: document.querySelector('body'),
+  };
+}
